通知リストのコンポーネントの切り出し
