+++ app/javascript/components/NotificationsBell/NotificationsBell.jsx
@@ -1,10 +1,8 @@
 import React, { useState } from 'react'
 import useSWR from 'swr'
 import BellButton from './BellButton'
+import Notifications from './Notifications'
 import fetcher from '../../fetcher'
-import dayjs from 'dayjs'
-import relativeTime from 'dayjs/plugin/relativeTime'
-dayjs.extend(relativeTime)
 
 export default function NotificationsBell() {
   const [showNotifications, setShowNotifications] = useState(false)
@@ -44,16 +42,7 @@ export default function NotificationsBell() {
             className="header-dropdown__background"
             onClick={clickOutsideNotifications}></label>
           <div className="header-dropdown__inner is-notification">
-            <ul className="header-dropdown__items">
-              {data.notifications.map((notification) => {
-                return (
-                  <Notification
-                    key={notification.id}
-                    notification={notification}
-                  />
-                )
-              })}
-            </ul>
+            <Notifications data={data} />
             <footer className="header-dropdown__footer">
               <a
                 href="/notifications?status=unread"
@@ -79,36 +68,5 @@ export default function NotificationsBell() {
         </div>
       )}
     </div>
-  )
-}
-
-function Notification({ notification }) {
-  const createdAtFromNow = (createdAt) => {
-    return dayjs(createdAt).fromNow()
-  }
-
-  return (
-    <li className="header-dropdown__item">
-      <a
-        href={notification.path}
-        className="header-dropdown__item-link unconfirmed_link">
-        <div className="header-notifications-item__body">
-          <span
-            className={`a-user-role header-notifications-item__user-icon is-${notification.sender.primary_role}`}>
-            <img
-              src={notification.sender.avatar_url}
-              className="a-user-icon"
-              alt="User Icon"
-            />
-          </span>
-          <div className="header-notifications-item__message">
-            <p className="test-notification-message">{notification.message}</p>
-          </div>
-          <time className="header-notifications-item__created-at">
-            {createdAtFromNow(notification.created_at)}
-          </time>
-        </div>
-      </a>
-    </li>
   )
 }
